backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
iniconfig==2.1.0
packaging==25.0
pillow==11.1.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.0.2
pytest-django==4.8.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
